fix: pathological graph traversal in findMergeBase
We encountered a situation where findMergeBase runs effectively forever even on a moderately sized repo. The problem is that there are many different paths to the same commits, and it permanently grows the `heads` list each time it finds an existing node via a new path, even under the same walker / `index`.

The fix was to never have the same walker visit the same commit more than once.

docs: add @ef4 as a contributor